+++ src/main/resources/release_notes.md
@@ -5,6 +5,9 @@ layout: page
 
 Changelist HO! 3.0
 
+## Squad
+  - Fix Sorting by Position #397
+
 ## League Details
  - Promotion / Demotion status displayed on week 14 and 15 of season. #247